Move DeriveStrategyThenCreate
